+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added new dialogues with Hermes. 
 
 Other fixes and improvements:
+- Started the addition of the Sisyphus liberation questline - you will be able to receive the new minor prophecy, but will not be able to complete it yet.
 - Updated a number of boon icons with enhanced visuals.
 - Fixed: When dying in Tartarus after room exits have been unlocked, part of the room reward preview animation is overlayed over the blacked-out death screen.
 - Fixed: Some dialogues are never eligible to play.